backend/requirements.txt
langchain>=0.3.0
langchain-ollama>=0.2.0
langgraph>=0.2.0
langchain-core>=0.3.0
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
mangum>=0.19.0
python-dotenv>=1.0.0
pydantic>=2.9.0
python-multipart>=0.0.12
